added main html, css and image files
Hi everyone! Here is a start. We need to do the about page and contact page. If you could write in a short description of each woman in STEM where it says 'description' that would be great (ctrl+f then search "description" in index.html to find where to write in there). Also anywhere where it says something like 'title here' or 'some text' are placeholders so feel free to write in there and tinker around with the code. Might also be worth adding hyperlinks for the social icons :)